LayoutTests/fast/css/hover-affects-ancestor-expected.txt

This test ensures that ancestor element hover rules are not affected when we hover its contained elements




http://www.webkit.org
Button test
Radio test
Checkbox test
testselect: PASS
testinput: PASS
testtextarea: PASS
testanchor: PASS
testbutton: PASS
testradio: PASS
testcheckbox: PASS
